// src/modules/v1/practice-writing/enums/kindQuestion.enum.ts
export enum KindQuestion {
    KIND_430002 = '430002',
    KIND_530002 = '530002',
    KIND_530003 = '530003',
    KIND_630001 = '630001',
}

export enum TypeRouteUserEnum {
    PRACTICE = 'practice',
    EVALUATE = 'evaluate',
    TEST = 'test',
}

export enum LevelHSKEnum {
    HSK1 = '1',
    HSK2 = '2',
    HSK3 = '3',
    HSK4 = '4',
    HSK5 = '5',
    HSK6 = '6',
    HSK7 = '7',
    HSK101 = '101',
    HSK102 = '102',
    HSK103 = '103',
}


export enum KindPracticeQuestionEnum {
    KIND_110001 = '110001',
    KIND_110002 = '110002',
    KIND_110003 = '110003',
    KIND_110004 = '110004',
    KIND_120001 = '120001',
    KIND_120002 = '120002',
    KIND_120003 = '120003',
    KIND_120004 = '120004',
    KIND_210001 = '210001',
    KIND_210002 = '210002',
    KIND_210003 = '210003',
    KIND_210004 = '210004',
    KIND_220001 = '220001',
    KIND_220002 = '220002',
    KIND_220003 = '220003',
    KIND_220004 = '220004',
    KIND_310001 = '310001',
    KIND_310002 = '310002',
    KIND_310003 = '310003',
    KIND_310004 = '310004',
    KIND_320001 = '320001',
    KIND_320002 = '320002',
    KIND_320003 = '320003',
    KIND_330001 = '330001',
    KIND_330002 = '330002',
    KIND_410001 = '410001',
    KIND_410002 = '410002',
    KIND_410003 = '410003',
    KIND_410003_1 = '410003_1',
    KIND_410003_2 = '410003_2',
    KIND_420001 = '420001',
    KIND_420002 = '420002',
    KIND_420003 = '420003',
    KIND_420003_1 = '420003_1',
    KIND_420003_2 = '420003_2',
    KIND_430001 = '430001',
    KIND_430002 = '430002',
    KIND_510001 = '510001',
    KIND_510002 = '510002',
    KIND_510002_1 = '510002_1',
    KIND_510002_2 = '510002_2',
    KIND_520001 = '520001',
    KIND_520002 = '520002',
    KIND_520003 = '520003',
    KIND_530001 = '530001',
    KIND_530002 = '530002',
    KIND_530003 = '530003',
    KIND_610001 = '610001',
    KIND_610002 = '610002',
    KIND_610003 = '610003',
    KIND_620001 = '620001',
    KIND_620002 = '620002',
    KIND_620003 = '620003',
    KIND_620004 = '620004',
    KIND_630001 = '630001',
    KIND_710001 = '710001',
    KIND_710002 = '710002',
    KIND_710003 = '710003',
    KIND_720001 = '720001',
    KIND_720002 = '720002',
    KIND_720003 = '720003',
    KIND_730001 = '730001',
    KIND_730002 = '730002',
    KIND_740001 = '740001',
    KIND_740002 = '740002',
    KIND_750001 = '750001',
    KIND_750002 = '750002',
    KIND_750003 = '750003',
    KIND_HSKKCC1 = 'HSKKCC1',
    KIND_HSKKCC2 = 'HSKKCC2',
    KIND_HSKKCC3 = 'HSKKCC3',
    KIND_HSKKSC1 = 'HSKKSC1',
    KIND_HSKKSC2 = 'HSKKSC2',
    KIND_HSKKSC3 = 'HSKKSC3',
    KIND_HSKKTC1 = 'HSKKTC1',
    KIND_HSKKTC2 = 'HSKKTC2',
    KIND_HSKKTC3 = 'HSKKTC3',
}